+++ strided/dcovmatmtk/test/test.dcovmatmtk.js
@@ -25,6 +25,8 @@
 var tape = require( 'tape' );
 var Float64Array = require( '@stdlib/array/float64' );
 var isAlmostEqualFloat64Array = require( '@stdlib/assert/is-almost-equal-float64array' );
+var isSameFloat64Array = require( '@stdlib/assert/is-same-float64array' );
+var filledarray = require( '@stdlib/array/filled' );
 var dcovmatmtk = require( './../lib/dcovmatmtk.js' );
 
 
@@ -36,6 +38,7 @@ var rmcols = require( './fixtures/row_major_columns.json' );
 var rmu = require( './fixtures/row_major_upper.json' );
 var rml = require( './fixtures/row_major_lower.json' );
 var rmsm = require( './fixtures/row_major_sm.json' );
+var rmic = require( './fixtures/row_major_invalid_correction.json' );
 
 var cm = require( './fixtures/column_major.json' );
 var cmc = require( './fixtures/column_major_corrected.json' );
@@ -43,6 +46,7 @@ var cmrows = require( './fixtures/column_major_rows.json' );
 var cmu = require( './fixtures/column_major_upper.json' );
 var cml = require( './fixtures/column_major_lower.json' );
 var cmsm = require( './fixtures/column_major_sm.json' );
+var cmic = require( './fixtures/column_major_invalid_correction.json' );
 
 
 // TESTS //
@@ -421,5 +425,49 @@ tape( 'the function supports specifying a stride for the vector of known means (
 	out = dcovmatmtk( data.order, data.orientation, data.uplo, data.M, data.N, data.correction, m, data.strideM, A, data.lda, B, data.ldb );
 	t.strictEqual( out, B, 'returns expected value' );
 	t.strictEqual( isAlmostEqualFloat64Array( out, expected, 200 ), true, 'returns expected value' );
+	t.end();
+});
+
+tape( 'the function sets elements in the output matrix to `NaN` if provided an invalid correction argument (row-major)', function test( t ) {
+	var expected;
+	var data;
+	var out;
+	var A;
+	var B;
+	var m;
+
+	data = rmic;
+
+	m = new Float64Array( data.means );
+	A = new Float64Array( data.A );
+	B = new Float64Array( data.B );
+
+	expected = filledarray( NaN, B.length, 'float64' );
+
+	out = dcovmatmtk( data.order, data.orientation, data.uplo, data.M, data.N, data.correction, m, data.strideM, A, data.lda, B, data.ldb );
+	t.strictEqual( out, B, 'returns expected value' );
+	t.strictEqual( isSameFloat64Array( out, expected ), true, 'returns expected value' );
+	t.end();
+});
+
+tape( 'the function sets elements in the output matrix to `NaN` if provided an invalid correction argument (column-major)', function test( t ) {
+	var expected;
+	var data;
+	var out;
+	var A;
+	var B;
+	var m;
+
+	data = cmic;
+
+	m = new Float64Array( data.means );
+	A = new Float64Array( data.A );
+	B = new Float64Array( data.B );
+
+	expected = filledarray( NaN, B.length, 'float64' );
+
+	out = dcovmatmtk( data.order, data.orientation, data.uplo, data.M, data.N, data.correction, m, data.strideM, A, data.lda, B, data.ldb );
+	t.strictEqual( out, B, 'returns expected value' );
+	t.strictEqual( isSameFloat64Array( out, expected ), true, 'returns expected value' );
 	t.end();
 });